+++ frontend/src/pages/preQuestion/PreQuestionEdit.tsx
@@ -6,8 +6,6 @@ import styled from 'styled-components';
 import useLevellog from 'hooks/useLevellog';
 import usePreQuestion from 'hooks/usePreQuestion';
 
-import Error from 'pages/exception/Error';
-
 import { MESSAGE, ROUTES_PATH } from 'constants/constants';
 
 import Button from 'components/@commons/Button';
